jcli: /api/v0/stake/rewards - remove output_format
